Module-1 Project-3/src/main/java/AnagramChecker.java/AnagramChecker.java
package main.java;

import java.util.Arrays;

public class AnagramChecker {

    public static boolean areAnagrams(String str1, String str2) {
        if (str1 == null || str2 == null) {
            return false;
        }

        // Remove all non-alphanumeric characters and convert to lowercase
        String cleanedStr1 = str1.replaceAll("[^a-zA-Z0-9]", "").toLowerCase();
        String cleanedStr2 = str2.replaceAll("[^a-zA-Z0-9]", "").toLowerCase();

        //If lengths don't match, they can't be anagrams
        if (cleanedStr1.length() != cleanedStr2.length()) {
            return false;
        }

        // Convert to char arrays, sort them, and compare
        char[] charArray1 = cleanedStr1.toCharArray();
        char[] charArray2 = cleanedStr2.toCharArray();
        Arrays.sort(charArray1);
        Arrays.sort(charArray2);

        return Arrays.equals(charArray1, charArray2);
    }
}